+++ utils/test/feedback/fizz_buzz_test.exs
@@ -1,0 +1,5 @@
+defmodule Utils.Feedback.FizzBuzzTest do
+  use Utils.FeedbackCase,
+    module: Utils.Feedback.FizzBuzz,
+    file: "../exercises/fizzbuzz.livemd"
+end
